Add specific imports for everything. Remove import all from package.. Split CAS22 into seperate files. Create thermal.py for shared cooling functions.
