+++ pyfecons/costing/mfe/CAS20.py
@@ -1,6 +1,6 @@
-from pyfecons import M_USD
 from pyfecons.inputs import Inputs
 from pyfecons.data import Data, TemplateProvider
+from pyfecons.units import M_USD
 
 CAS_200000_TEX = 'CAS200000.tex'
