+++ pyfecons/costing/mfe/CAS25.py
@@ -1,7 +1,7 @@
-from pyfecons import M_USD
 from pyfecons.costing.calculations.conversions import inflation_factor_2019_2024
 from pyfecons.inputs import Inputs
 from pyfecons.data import Data, TemplateProvider
+from pyfecons.units import M_USD
 
 CAS_250000_TEX = 'CAS250000.tex'
